Add fleet life to each player
