+++ src/Game.py
@@ -1,8 +1,8 @@
 from copy import deepcopy
 from Config import getPlayersConfig
-from Board import initBoard, getBoardFromConfig, displayBoard, shoot
+from Board import initBoard, getBoardFromConfig, shoot
 from random import randint
-from constants import PLAYER_1, PLAYER_2
+from constants import PLAYER_1, PLAYER_2, FLEET_LIFE
 
 def switchPlayer(currentPlayer):
     if currentPlayer == PLAYER_1:
@@ -19,10 +19,12 @@ def initGame(config1, config2, size):
         PLAYER_1: {
             "primary": getBoardFromConfig(player1Config),
             "opponent": initBoard(10),
+            "life": FLEET_LIFE
         },
         PLAYER_2: {
             "primary": getBoardFromConfig(player2Config),
             "opponent": initBoard(10),
+            "life": FLEET_LIFE
         }
     }
 
@@ -33,3 +35,13 @@ def initGame(config1, config2, size):
         currentPlayer = PLAYER_2
 
     return boards, currentPlayer
+
+def decrementTargetFleetLife(boards, attacker):
+    boardsCopy = deepcopy(boards)
+
+    if attacker == PLAYER_1:
+        boardsCopy[PLAYER_2]["life"] -= 1
+    else:
+        boardsCopy[PLAYER_1]["life"] -= 1
+
+    return boardsCopy
