+++ src/Game_test.py
@@ -1,0 +1,6 @@
+import unittest
+
+class TestGameMethods(unittest.TestCase):
+
+    def test_game_init(self):
+        self.assertTrue(1)
